% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomocnicze_tfn.R
\name{.defuzyfikuj_macierz_tfn}
\alias{.defuzyfikuj_macierz_tfn}
\title{Defuzyfikacja macierzy TFN}
\usage{
.defuzyfikuj_macierz_tfn(macierz_decyzyjna)
}
\description{
Zamienia macierz (l, m, u) na macierz ostrą metodą (l + 4m + u) / 6.
}
\keyword{internal}
