% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomocnicze_tfn.R
\name{.defuzyfikuj_tfn}
\alias{.defuzyfikuj_tfn}
\title{Defuzyfikacja trójkątnej liczby rozmytej}
\usage{
.defuzyfikuj_tfn(l, m, u)
}
\description{
Stosuje ważony środek TFN: (l + 4m + u) / 6.
}
\keyword{internal}
