% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomocnicze_tfn.R
\name{.liczba_kryteriow}
\alias{.liczba_kryteriow}
\title{Liczba kryteriów w macierzy TFN}
\usage{
.liczba_kryteriow(macierz_decyzyjna)
}
\description{
Liczba kryteriów w macierzy TFN
}
\keyword{internal}
