% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagi_obiektywne.R
\name{.mapuj_metode_wag}
\alias{.mapuj_metode_wag}
\title{Mapowanie angielskich nazw metod wag}
\usage{
.mapuj_metode_wag(metoda_wag)
}
\description{
Mapowanie angielskich nazw metod wag
}
\keyword{internal}
