% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomocnicze_tfn.R
\name{.nazwy_alternatyw}
\alias{.nazwy_alternatyw}
\title{Nazwy alternatyw z macierzy decyzyjnej}
\usage{
.nazwy_alternatyw(macierz_decyzyjna)
}
\description{
Nazwy alternatyw z macierzy decyzyjnej
}
\keyword{internal}
