% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomocnicze_tfn.R
\name{.normalizuj_macierz_obiektywna}
\alias{.normalizuj_macierz_obiektywna}
\title{Macierz zorientowana dla obiektywnych metod wag}
\usage{
.normalizuj_macierz_obiektywna(macierz_ostra, typy_kryteriow)
}
\description{
Przekształca kryteria kosztowe i zyskowe tak, aby większa wartość oznaczała korzystniejszą ocenę.
}
\keyword{internal}
