% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomocnicze_tfn.R
\name{.normalizuj_min_max}
\alias{.normalizuj_min_max}
\title{Bezpieczna normalizacja do przedziału [0, 1]}
\usage{
.normalizuj_min_max(x, kierunek = "max")
}
\description{
Bezpieczna normalizacja do przedziału [0, 1]
}
\keyword{internal}
