% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagi_obiektywne.R
\name{.pobierz_finalne_wagi}
\alias{.pobierz_finalne_wagi}
\title{Ustalenie wag dla metod MCDA}
\usage{
.pobierz_finalne_wagi(
  macierz_decyzyjna,
  typy_kryteriow,
  metoda_wag = c("critic", "entropia"),
  wagi = NULL
)
}
\description{
Pobiera wagi obiektywne albo normalizuje wcześniej obliczony wektor wag.
}
\keyword{internal}
