% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomocnicze_tfn.R
\name{.rozszerz_typy_kryteriow}
\alias{.rozszerz_typy_kryteriow}
\title{Rozszerzenie typów kryteriów do kolumn TFN}
\usage{
.rozszerz_typy_kryteriow(typy_kryteriow, n_kryteriow)
}
\description{
Rozszerzenie typów kryteriów do kolumn TFN
}
\keyword{internal}
