% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{.skaluj_do_1_9}
\alias{.skaluj_do_1_9}
\title{Wewnętrzne skalowanie do przedziału 1-9}
\usage{
.skaluj_do_1_9(wektor)
}
\description{
Przekształca dowolną skalę (np. Likert 1-5, wartości ciągłe)
na wspólną skalę 1-9 używaną w macierzy decyzyjnej.
}
\keyword{internal}
