% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomocnicze_tfn.R
\name{.sprawdz_macierz_tfn}
\alias{.sprawdz_macierz_tfn}
\title{Walidacja rozmytej macierzy decyzyjnej}
\usage{
.sprawdz_macierz_tfn(macierz_decyzyjna)
}
\description{
Sprawdza podstawową strukturę macierzy TFN.
}
\keyword{internal}
