% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomocnicze_tfn.R
\name{.sprawdz_typy_kryteriow}
\alias{.sprawdz_typy_kryteriow}
\title{Walidacja typów kryteriów}
\usage{
.sprawdz_typy_kryteriow(typy_kryteriow, n_kryteriow)
}
\description{
Walidacja typów kryteriów
}
\keyword{internal}
