% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja.R
\name{.utworz_tabele_apa}
\alias{.utworz_tabele_apa}
\title{Pomocnicze formatowanie tabel APA}
\usage{
.utworz_tabele_apa(df, numer_tabeli, tytul, uwaga)
}
\description{
Pomocnicze formatowanie tabel APA
}
\keyword{internal}
