% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ranking.R
\name{fuzzy_meta_ranking}
\alias{fuzzy_meta_ranking}
\title{Rozmyty Meta-Ranking}
\usage{
fuzzy_meta_ranking(
  decision_mat,
  criteria_types,
  weights = NULL,
  bwm_best = NULL,
  bwm_worst = NULL,
  lambda = 0.5,
  v = 0.5
)
}
\description{
Agreguje wyniki z wielu metod MCDA w jeden ranking.
}
