% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja.R
\name{plot.fuzzy_vikor_res}
\alias{plot.fuzzy_vikor_res}
\title{Wykres bąbelkowy VIKOR dla aliasu angielskiego}
\usage{
\method{plot}{fuzzy_vikor_res}(x, ...)
}
\arguments{
\item{x}{Obiekt klasy `fuzzy_vikor_res`.}

\item{...}{Dodatkowe argumenty przekazywane do wykresu VIKOR.}
}
\value{
Obiekt `ggplot`.
}
\description{
Metoda zgodności dla obiektów zwracanych przez `fuzzy_vikor()`.
}
