% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja.R
\name{tabela_apa}
\alias{tabela_apa}
\title{Generowanie tabeli APA}
\usage{
tabela_apa(x, tytul = NULL, ...)
}
\arguments{
\item{x}{Obiekt wynikowy z funkcji pakietu, np. `rozmyty_topsis_wynik`.}

\item{tytul}{Opcjonalny tytul tabeli.}

\item{...}{Dodatkowe argumenty przekazywane do metod S3.}
}
\value{
Obiekt klasy `flextable`.
}
\description{
Funkcja przeksztalca wyniki analizy MCDA (TOPSIS, VIKOR, WASPAS,
Meta-Ranking) w sformatowana tabele zgodna ze stylem APA, gotowa
do druku albo zapisu do dokumentu Word.
}
